requirements.txt
boto3
pandas
streamlit==1.23.0
botocore==1.34.107
mysql-connector-python 
pyodbc